react helmet added - reviewed by @Andras92
