+++ frontend/package.json
@@ -20,11 +20,12 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
-    "react-use": "^17.2.4",
+    "react-helmet-async": "^1.0.9",
     "react-icons": "^4.2.0",
     "react-leaflet": "^3.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-use": "^17.2.4",
     "styled": "^1.0.0",
     "styled-components": "^5.3.0",
     "swiper": "^6.7.0",
