+++ frontend/src/Map/Map.js
@@ -2,9 +2,19 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import Markers from './Markers';
 import 'leaflet/dist/leaflet.css';
+import { Marker, TileLayer, MapContainer } from 'react-leaflet';
+import styled from 'styled-components';
+import { useGeolocation } from 'react-use';
+import UserIcon from '../Icon/UserIcon';
 
 export default function Map() {
+  const myPosition = useGeolocation();
   const [coordinates, setCoordinates] = useState([]);
+
+  const StyledMapContainer = styled(MapContainer)`
+    width: 100%;
+    height: 80vh;
+  `;
 
   useEffect(() => {
     axios
@@ -18,8 +28,25 @@ export default function Map() {
   }, []);
 
   return (
-    <div>
+    <StyledMapContainer
+      center={[47.497913, 19.040236]}
+      zoom={10}
+      scrollWheelZoom={true}
+    >
+      <TileLayer
+        attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+        url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+      />
       <Markers coordinates={coordinates} />
-    </div>
+      {myPosition.latitude !== null && (
+        <>
+          <pre>{(myPosition, null, 2)}</pre>
+          <Marker
+            icon={UserIcon}
+            position={[myPosition.latitude, myPosition.longitude]}
+          ></Marker>
+        </>
+      )}
+    </StyledMapContainer>
   );
 }
